Split POSIX enumeration into individual modules
There are only a few variants but many imports, functions, ... had
individual annotations for conditional compilation. Lets split more
coarsely and have less noise in the individual variants.
