+++ src/posix/mod.rs
@@ -1,7 +1,3 @@
-pub use self::enumerate::*;
-pub use self::serialportext::SerialPortExt;
-pub use self::tty::*;
-
 mod enumerate;
 mod error;
 mod flock;
@@ -10,3 +6,7 @@ mod poll;
 mod serialportext;
 mod termios;
 mod tty;
+
+pub use self::enumerate::available_ports;
+pub use self::serialportext::SerialPortExt;
+pub use self::tty::*;
